Remove state support for legacy leases. Extract the lease upgrade related logic into its own file so all the remaining legacy lease stuff is kept together.
