--- state/lease/package_test.go
@@ -1,14 +1,0 @@
-// Copyright 2015 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package lease_test
-
-import (
-	stdtesting "testing"
-
-	coretesting "github.com/juju/juju/testing"
-)
-
-func TestPackage(t *stdtesting.T) {
-	coretesting.MgoTestPackage(t)
-}
